Extract RangeList
