+++ FastAerialTrainer/FastAerialTrainer.vcxproj
@@ -111,6 +111,7 @@
     <ClCompile Include="FastAerialTrainer.cpp" />
     <ClCompile Include="FastAerialTrainerGUI.cpp" />
     <ClCompile Include="PersistentStorage.cpp" />
+    <ClCompile Include="RangeList.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="GuiBase.h" />
@@ -132,6 +133,7 @@
     <ClInclude Include="pch.h" />
     <ClInclude Include="FastAerialTrainer.h" />
     <ClInclude Include="PersistentStorage.h" />
+    <ClInclude Include="RangeList.h" />
     <ClInclude Include="version.h" />
   </ItemGroup>
   <ItemGroup>
